update sentence count regex to use word boundries
fixes #139. this removes the spliting on a match, and created a regex
that defines a sentence.

this regex is not complete, and I'm choosing not to commit a more
complete one. a better regex would account for abreviations, like
"T.V.", but fixing the long standing bug would cause more issues now.